main: Clean up main configuration parsing code
Make the format of the configuration file more robust without stepping
on many parts of the program.  However, replacing the string
"NonBoussinesq" with "NON_BOUSSINESQ" all over the program can be
justified by the fact that the string "NonBoussinesq" appears in a
function name, and this can be confusing for the reader and editor
when performing a replacement.

Also cleanup some of the files directly related to the main program so
that the obnoxious license header, extra whitespace, and verbose
comments don't distract the reader from the actual code.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>